requirements.txt
Django==3.0.1
djangorestframework==3.10.0

python-decouple==3.3
gunicorn==20.0.4
django-heroku==0.3.1
psycopg2==2.8.4

flake8==3.7.0
requests==2.22.0

# django-rest-auth==0.9.5
# django-allauth==0.40.0
